requirements.txt
bpython==0.22.1
brotli==1.0.9
ConfigParser==5.2.0
cryptography==2.8
cx_Oracle==8.3.0
Cython==0.29.28
django_daraja==1.2.0
django_debug_toolbar==3.2.4
dl==0.1.0
docutils==0.18.1
HTMLParser==0.0.2
importlib_metadata==1.5.0
importlib_resources==5.4.0
ipdb==0.13.9
ipython==8.1.1
jnius==1.1.0
keyring==18.0.1
lockfile==0.12.2
lxml==4.7.1
MySQL-python==1.2.5
numpy==1.17.4
ordereddict==1.1
Pillow==9.0.1
protobuf==3.19.4
psycopg2==2.9.3
pylibmc==1.6.1
pymemcache==3.5.1
pyOpenSSL==22.0.0
python_memcached==1.59
pytz==2019.3
pywatchman==1.4.1
PyYAML==6.0
tblib==1.7.0
toml==0.10.2
tornado==5.1.1
typing_extensions==4.1.1
xmlrpclib==1.0.1
zipp==1.0.0
